Work on serialization, settings, and more
Currently trying to fix a memory bug related to fonts.

- Updated DirectXTK to latest
- Work on settings interface
- Fixed DEVICE_DRAW_RESOURCE_FORMAT_SAMPLE_UNSUPPORTED when drawing a Model3D with no texture loaded

+++ 3DRadSpace/Engine3DRadSpace/Math/Math.cpp
@@ -61,7 +61,7 @@ bool DLLEXPORT Engine3DRadSpace::Math::AABB(double p1, double w1, double p2, dou
 
 double DLLEXPORT Engine3DRadSpace::Math::SmoothStep(double value1, double value2, double amount)
 {
-    float result = std::clamp(amount, 0.0, 1.0);
+    double result = std::clamp(amount, 0.0, 1.0);
     result = Hermite(value1, 0, value2, 0, result);
 
     return result;
